use cargo-release alpha version numbers if available
When you use cargo-release to update the version number, the next
working version winds up being something like 0.1.2-alpha.0. That is not
a valid RPM version number, though. However, you can split the version
number along the - and take the latter half as part of the release. This
change will set the version number for the RPM to be (from the example)
0.1.2 with a release of 0.alpha.0. If there is no release found in the
version number, assume a release of 1 (which will be later than the
pre-release version numbers provided by cargo-release). Fixes #3.